Fix module build for kernel 6.13.
I don't really understand why this works. The old makefile used paths
relative to the current directory (e.g. `./src/tuxedo_keyboard.o`), and
has worked for several prior revisions. However, after updating to
the 6.13 kernel, it seems the build isn't recognized as a module, and
when modpost runs, it produces errors like
	`ERROR: modpost: missing MODULE_LICENSE() in src/tuxedo_keyboard.o`
Removing the './' prefix from the makefile appears to fix that. Did the
assumed directory change for the module build?